DEV : Added ProjectModel, ProjectFactory
