+++ app/Models/Page.php
@@ -3,9 +3,8 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
 
-class Page extends Model
+class Page extends ProjectModel
 {
     use HasFactory;
 }
