+++ app/Models/PageHit.php
@@ -3,9 +3,8 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
 
-class PageHit extends Model
+class PageHit extends ProjectModel
 {
     use HasFactory;
