+++ database/factories/PageFactory.php
@@ -2,13 +2,12 @@
 
 namespace Database\Factories;
 
-use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 
 /**
  * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Page>
  */
-class PageFactory extends Factory
+class PageFactory extends ProjectFactory
 {
     /**
      * Define the model's default state.
